kvs2: remove db pointer from nodeptr
prior to this commit every node pointer carried around a pointer to its
associated database so that a pointer could be resolved. it needed the
db pointer so it could access the correct cache/log/etc... an extra two
8 byte pointers per node adds up.

this patch stashes a current db pointer in thread local storage, and
sets this pointer any time a thread starts to interact with the database
(e.g. transaction or iterator), and clears the variable before
returning.

we use a basic raii technique to set and clear this, but it is extremely
important that its always set and cleared.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>